Implemented appointment details page
The appointment details page for lawyers has the ability to confirm or reject an appointment request, that has been requested to them. The dialog box allows the setting of the timestamp if it is not already set.
Decreased the connectionLimit for the MariaDB connector to 4 since it can't create more than 5-6 connections for some reason.
The appointments table has a constraint IfStatusConfirmedThenTimestampIsNotNullCheck, which checks that the timestamp is set by the time an appointment is confirmed.