Added MIME type checking for the photos to upload
Added MIME type checking for the photos to upload during registration.
The MIME type checking is done on both the client and server side. It is more thorough on the server side as it uses the package file-type.

Sets white-space: normal, for v-card-title, i.e. text on card titles is not sliced off when overflowing on smaller screen sizes.