Fixed small bug.
Fixed small bug where the case documents page would not refresh the masonry when the documents list is changed.
Added :key for messages and message groups by day.
Changes the background color of the CaseDocument based on who uploaded it, the color scheme is identical to the one used in the chat group.